+++ src/Aks/Aks/Commands/NewAzureRmAks.cs
@@ -142,6 +142,9 @@ namespace Microsoft.Azure.Commands.Aks
             HelpMessage = "Availability zones for cluster. Must use VirtualMachineScaleSets AgentPoolType.")]
         public string[] AvailabilityZone { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "The resource group containing agent pool.")]
+        public string NodeResourceGroup { get; set; }
+
         private AcsServicePrincipal acsServicePrincipal;
 
         public override void ExecuteCmdlet()
@@ -350,6 +353,7 @@ namespace Microsoft.Azure.Commands.Aks
                 tags: TagsConversionHelper.CreateTagDictionary(Tag, true),
                 dnsPrefix: DnsNamePrefix,
                 kubernetesVersion: KubernetesVersion,
+                nodeResourceGroup: NodeResourceGroup,
                 agentPoolProfiles: new List<ManagedClusterAgentPoolProfile> { defaultAgentPoolProfile },
                 linuxProfile: linuxProfile,
                 windowsProfile: windowsProfile,