+++ src/Network/Network/Models/NetworkManager/PSNetworkManagerStaticMember.cs
@@ -1,0 +1,14 @@
+﻿using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Azure.Commands.Network.Models.NetworkManager
+{
+    public class PSNetworkManagerStaticMember : PSNetworkManagerBaseResource
+    {
+        public string NetworkManagerName { get; set; }
+        public string NetworkGroupName { get; set; }
+        public string ResourceId { get; set; }
+    }
+}