+++ src/Accounts/Accounts/ChangeLog.md
@@ -20,6 +20,10 @@
 
 ## Upcoming Release
 
+## Version 2.9.1
+* Implemented `SupportsShouldProcess` for `Invoke-AzRestMethod`
+* Supported giving suggestions if an Azure PowerShell command cannot be found, for example when there is a typo in command name.
+
 ## Version 2.9.0
 * Supported exporting and importing configurations by `Export-AzConfig` and `Import-AzConfig`.
 * Fixed an issue that Az.Accounts may fail to be imported in parallel PowerShell processes. [#18321]