+++ src/Batch/Batch/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed a bug wherein creating a new JobSchedule does not properly submit Output Files.
 
 ## Version 3.2.0