+++ src/DnsResolver/Changelog.md
@@ -19,6 +19,10 @@
 -->
 ## Upcoming Release
 
+## Version 0.2.1
+* Reformatted exception message for cmdlets
+* Upgraded API version to 2022-07-01 
+
 ## Version 0.2.0
 * Preview release for module Az.DnsResolver